Review GenericTimer: 32-bit ARR on TIM2/TIM5
CalculatePeriod() returned uint16_t and clamped ARR to UINT16_MAX,
artificially flooring the achievable frequency at ~0.153 Hz on every
instance. TIM2 and TIM5 are 32-bit on the F4 (RM0090 17.3.1), so they
can hold a far larger ARR. Widened the return to uint32_t and clamp
per timer width: 0xFFFFFFFF for TIM2/TIM5, 0xFFFF (unchanged) for the
16-bit instances. mHandle.Init.Period is already uint32_t so the
value now flows through unclamped on the 32-bit timers.

Behaviour for 16-bit timers is identical; additive, non-breaking.
All five ARM Debug targets build clean. Closes driver_update.md
GenericTimer #2.

+++ drivers/drivers/GenericTimer/GenericTimer.hpp
@@ -129,7 +129,7 @@ private:
     void CheckAndEnablePeripheralClock(const GenericTimerInstance& instance);
     void CheckAndDisablePeripheralClock(const GenericTimerInstance& instance);
     uint32_t GetTimerInputClockFreq(const GenericTimerInstance& instance);
-    uint16_t CalculatePeriod(float desiredFrequency);
+    uint32_t CalculatePeriod(float desiredFrequency);
     IRQn_Type GetIRQn(const GenericTimerInstance& instance);
     TimerIRQ::Slot GetSlot(const GenericTimerInstance& instance);
     void SetIRQn(IRQn_Type type, uint32_t preemptPrio, uint32_t subPrio);
